requirements.txt
django==5.2.3
djangorestframework==3.16.0
django-jazzmin==2.6.0
pillow==10.3.0
drf-spectacular==0.27.2
django-filter==24.3
fcm-django==2.1.0